Update get_klifs_residues_mol2topdb
Be cleaner in generating the residue_ids from the mol2 file for the lookup in the PDB file.

Before ID generated from everything after first three letters (mostly PDB ID) in subst_name of mol2 file - does not work always, e.g. CA134 refers to element and ID is 134 and not 34.